requirements.txt
anthropic>=0.30.0
requests>=2.32.0
python-dotenv>=1.0.0
google-auth>=2.30.0
google-api-python-client>=2.130.0
